Add FIND_PACKAGE call for BLAS
* The CMAKE system was failing to add BLAS_LIBRARIES to the linked
libraries list used during the final build stages. This commit adds a
call to FIND_PACKAGE for BLAS and adds the BLAS_LIBRARIES to the
librarie list to link during the final build stages. This cleans up an
issue with installing hopr when nonstandard paths are used for BLAS
insallations